+++ src/components/LogoutButton.tsx
@@ -1,23 +1,13 @@
 // src/components/LogoutButton.tsx
 import React from "react";
-import { useNavigate } from "react-router-dom";
-import { queryClient } from "@src/main";
+import { useLogout } from "@src/hooks/useLogout";
 
 export default function LogoutButton() {
-  const nav = useNavigate();
-
-  const logout = () => {
-    localStorage.removeItem("accessToken");
-    localStorage.removeItem("refreshToken");
-
-    queryClient.invalidateQueries({ queryKey: ["me"] });
-
-    nav("/", { replace: true });
-    window.dispatchEvent(new StorageEvent("storage")); 
-  };
+  const { logout } = useLogout();
 
   return (
     <button
+      type="button"
       onClick={logout}
       className="text-[14px] text-[var(--Gray20)] hover:text-[var(--Black)]"
     >
